% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTArun.R
\name{CTArun}
\alias{CTArun}
\title{Execute an CTA model run.}
\usage{
CTArun(
  run = "",
  path = getwd(),
  data = "data.txt",
  out = "model1.txt",
  vstart = "",
  vend = "",
  class = "",
  attribute = "",
  categorical = "",
  include = "",
  exclude = "",
  direction = "",
  force = F,
  forcenode = "",
  nodevar = "",
  miss = "",
  weight = "",
  usefisher = F,
  mciter = "",
  cutoff = "",
  stop = "",
  nopriors = F,
  mindenom = "",
  maxlevel = "",
  prune = "",
  skipnode = "",
  enumerate = "",
  loo = "",
  overwrite = FALSE
)
}
\arguments{
\item{run}{A numerical value specifying the run folder containing the data}

\item{path}{The working directory of the project stored as \code{path}}

\item{data}{A character name specifying the data.txt file in the model folder}

\item{out}{A character name specifying the output file with "model.out" as
the default}

\item{vstart}{A character name specifying the start variable, see \code{key}
object from \code{\link{ODAload}}}

\item{vend}{A character name specifying the end variable, see \code{key}
object from \code{\link{ODAload}}}

\item{class}{A character name specifying the class variable, see \code{key}
object from \code{\link{ODAload}}}

\item{attribute}{A character name specifying the attributes, see \code{key}
object from \code{\link{ODAload}}}

\item{categorical}{An optional character name specifying categorical
attributes, see \code{key} object from \code{\link{ODAload}}}

\item{include}{An optional character name specifying the variable and a value
of observations that are included e.g., \code{"v2=2"} or \code{"v3>=50"}}

\item{exclude}{An optional character name specifying the variable and a value
that are excluded}

\item{direction}{An optional character name and direction e.g., \code{v2 < 1
or v2 LT 1} specifying a directional hypothesis for the class variable}

\item{force}{A logical specifying whether CTA should insert \code{nodevar}
attribute at \code{forcenode} node in the solution tree}

\item{forcenode}{A numeric specifying the node to insert \code{nodevar} if
\code{force} = T}

\item{nodevar}{A character string specifiying the variable to be inserted at
node \code{forcenode} if \code{force} = T}

\item{miss}{A numeric value specifying a missing or \code{NA} value in the
data with a default value set at \code{-9}, see \code{\link{ODAclean}}}

\item{weight}{An optional character name of a variable containing a positive,
non zero, weight value; cannot be the same as variables declared as
\code{class} \code{attribute} or \code{categorical} attributes}

\item{usefisher}{A logical specifying that all p values for categorical
attributes will be determined using Fisher's Exact test rather than Monte
Carlo}

\item{mciter}{A numeric value specifying the number of Monte Carlo iterations
used to estimate the type I error rate for a given attribute within the
solution tree}

\item{cutoff}{A cutoff value for the Type I error rate as an additional
argument to \code{mciter} and \code{stop}}

\item{stop}{A confidence level in percent which will stop the processing of a
given attribute after the number of Monte Carlo iterations specified in
\code{mciter} are completed and the Type I error \code{cutoff} is exceeded}

\item{nopriors}{A logical specifying whether the ODA criterion will be
weighted by the reciprocal of sample class membership. The default of
leaving 'PRIORS' undeclared is equivalent to \code{nopriors} = F}

\item{mindenom}{A numeric value that specifies that only those attributes
which yield a solution tree node denominator of size 'MINDENOM' or larger
will be allowed in the solution tree}

\item{maxlevel}{A numeric value that specifies the deepest level of the
solution tree}

\item{prune}{A numeric value specifying the upper bound Type I error rate
with which to prune the solution tree or. When \code{nopriors} = T, then
\code{prune} should be specified as 'NOPRIORS'. The default value is 0.05}

\item{skipnode}{A numeric value specifying the node number to skip.}

\item{enumerate}{A logical specifying whether the top three attribute or a
limited character string. 'ROOT' specifies that only the top node will have
all attributes evaluated. 'MINOBS' followed by a numeric value e.g.,
'MINOBS 1' which will only allow solution trees with at least MINOBS.}

\item{loo}{A character string or numeric value specifying the upper bound
Type 1 error rate required in jackknife i.e., leave-one-out or LOO cross
generalizability analysis for every attribute in the tree. A specification
of 'STABLE', the default value, requires that attributes have the same ESS
in training and in LOO in order to be included in the solution tree}

\item{overwrite}{An optional logical value specifying whether previous files
in the model folder should be overwritten with \code{overwrite = FALSE} as
the default}
}
\value{
Nothing is returned. Three files are created in the model folder:
  \item{MODEL.out}{The model file that contains the commands from CTA syntax
  and analysis results, see \code{\link{ODAmanual}}. This file is required
  for \code{\link{CTAparse}}.}\item{cta.pgm}{The command file that contains
  all of the commands for CTA passed from R based on user input to
  \code{CTArun}.}\item{ctarun.bat}{The batch file that must be executed to
  begin the model run.}
}
\description{
Creates and processes files needed for CTA.exe program
}
\details{
The working directory of the project is stored as \code{path}. All
  files needed for the model run must be located in the appropriate model run
  folder. See \code{\link{ODAclean}}.

  \code{CTArun} will produce a command file with a .pgm extension, a model
  file with a .out extension, and a batch file with .bat extension. The .bat
  file must be executed within the model directory. Model output files can be
  parsed using \code{\link{CTAparse}}.
}
\examples{
# Not run:
# CTArun(run=1,vstart="v1", vend="v201", class="v46 to v201", exclude ="v3=1", attr="v2 v4 to v44", mindenom = 1, enumerate = T, loo="STABLE")

}
\references{
Yarnold P.R. and Soltysik R.C. (2005). \emph{Optimal data
  analysis: Guidebook with software for Windows}. APA Books.

  Yarnold, P.R. and Soltysik, R.C. (2016). \emph{Maximizing Predictive
  Accuracy}. ODA Books. DOI: 10.13140/RG.2.1.1368.3286
}
\seealso{
\code{\link{ODAmanual}} \code{\link{ODAclean}} \code{\link{ODAload}}
}
\author{
Nathaniel J. Rhodes
}
