% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{.onLoad getOption package settings}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{defunct}

\item{pkgname}{defunct}
}
\value{
invisible()
}
\description{
.onLoad getOption package settings
}
