% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOVOboot.R
\name{NOVOboot}
\alias{NOVOboot}
\title{Perform novometric bootstrap analysis using a classification model}
\usage{
NOVOboot(
  data = "",
  run = "",
  predictor = "",
  outcome = "",
  nboot = "",
  seed = ""
)
}
\arguments{
\item{data}{The name of a valid object created by \code{\link{ODAparse}} that
begins with \code{oda.list.x} where \code{x} is the \code{run}.}

\item{run}{The run number of the ODA model created using \code{\link{ODArun}}
and parsed using \code{\link{ODAparse}}.}

\item{predictor}{The model number within the \code{\link{ODArun}} which is
read using \code{\link{ODAparse}}. This is the model that is compared to
chance.}

\item{outcome}{The outcome number, when more than one outcome was evaluated.
For example, if three outcomes were evaluated simultaneously and the impact
of outcome one was of interest, then enter 1.}

\item{nboot}{The number of bootstrap replicates. Both model and chance will
be evaluated using this number of replicates using a 50\% resampling with
replacement. The default value is 25,000 replicates.}

\item{seed}{The seed number passed to \code{\link{set.seed}} that serves as
the origin of the pseudorandom numbers generated for the bootstrap
resampling. The default seed number is the current system time.}
}
\value{
An array of percentiles from 0\% to 100\% capturing the resampled
  model and chance performance metrics. The simulated bootstraps are stored
  within the object \code{novo.boot.x} which is appended with the current
  \code{run} number \code{x}.
}
\description{
Perform novometric bootstrap analysis using a classification model
}
\details{
The first axiom of novometric theory states that, for a random
  statistical sample "S" consisting of a class variable, one or more
  attributes, and a weight, the corresponding exact discrete confidence
  intervals, CIs, for model and for chance do NOT overlap. That is, a
  signficant model exists. If the CIs overlap, the effect is not
  statistically significant. However, if the CIs do not overlap, then the
  effect strength of the model is statistically significant at the confidence
  level selected by the user.

  \code{NOVOboot} reports the distribution of ESS in quantiles from 0\% to
  100\%. This provides a mechanism to evaluate the validity of an ODA model
  through bootstrap resampling with replacement. The same methodology can be
  applied to any classifcation model wherein a two class and two attribute
  confusion matrix can be formulated. It is axiomatic that the exact discrete
  confidence intervals for model and chance define the boundry between
  "signal" and "noise". A distribution of p values from Fisher's Exact tests
  conducted for each bootstrap replicate are also supplied and can be
  graphically evaluated, as in the example below.
}
\examples{
## Not run
## NOVOboot(data=oda.list.1,run=1,predictor=8,outcome=1,seed=1234)

## Example of a moderate effect size (ESS) Novometric bootstrap confidence interval
ess <- ((((14/14)+(20/64))/2)-0.5)/.5        # 31.25\% ESS, a moderate effect
data <- matrix(c(20,0,44,14),ncol=2,nrow=2,dimnames=list(c(0,1),c(0,1)))
data.raw <- epitools::expand.table(data)
data.tab <- list(table(cbind(data.raw[1],data.raw[2]),dnn=c("v1","x")))
oda.list.1 <- list()                         # supply list formatted for NOVOboot
oda.list.1[[1]] <- do.call("list",data.tab)  # supply data for NOVOboot

# NOVOboot(data=oda.list.1,run=1,predictor=1,outcome=1,seed=1234, nboot=25000)
# hist(novo.boot.1$p, main="Distribution of P-values \n by Fisher's Exact Test")

}
\references{
Yarnold, P.R. (2020). Reformulating the First Axiom of Novometric
  Theory: Assessing Minimum Sample Size in Experimental Desig \emph{Optimal
  Data Analysis} \bold{9}, 7-8.
  \url{https://odajournal.files.wordpress.com/2020/01/v9a2.pdf}

  Yarnold, P.R. and Soltysik, R.C. (2016). \emph{Maximizing Predictive
  Accuracy}. ODA Books. DOI: 10.13140/RG.2.1.1368.3286
}
\seealso{
\code{\link{ODArun}} \code{\link{ODAparse}} \code{\link{set.seed}}
}
