% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAload.R
\name{ODAload}
\alias{ODAload}
\title{Load data files and variable key for ODA}
\usage{
ODAload(run = "", path = getwd(), ...)
}
\arguments{
\item{run}{The numerical value of the folder number containing the data
specified by the run number. This number will also be used to name objects
uniquely by appending \code{run} to the object e.g. data.1 for Run 1. At
least one value for this parameter must be supplied by the user as no
defaults are supplied.}

\item{path}{The working directory of the project stored. the working
directory should be set above the level of the Runs folder.}

\item{...}{Additional data files to load and review, if desired.}
}
\value{
The following objects are loaded into R \item{data}{Data frame of
  data.csv from specified \code{run} folder.} \item{key}{Data frame
  containing 2 columns: the variable names from the \code{data} and an
  ODA-friendly alias e.g., v1 v2}
}
\description{
Loads the primary data file from the specified run folder and
  generates an ODA-friendly key for the user.
}
\details{
The current working directory is stored as the \code{path} and the
  files to be loaded must be located in the \code{Run} folder.
}
\examples{
# Not run:
# ODAload(1)
}
\author{
Nathaniel J. Rhodes
}
