% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAparse.R
\name{ODAparse}
\alias{ODAparse}
\title{Parse ODA output files.}
\usage{
ODAparse(run = "", ...)
}
\arguments{
\item{run}{The numerical value of the run folder in which the data file is
stored}

\item{...}{Additional run numbers specified as a list}
}
\value{
The following objects with the run number appended are returned for
  ODA models: \item{oda.data}{Data frame based on the data.csv file from
  specified \code{run} folder.} \item{oda.key}{Data frame containing 2
  columns: the variable names from the \code{oda.data} and an ODA friendly
  alias e.g., v1 v2 for each attribute variable included in the ODA model.}
  \item{oda.list}{Data frame containing the confusion matrix for each ODA
  model contained in the model.out file. This output can be used to evaluate
  the reproducibility of the model results using a Novometric bootstrap
  analysis, see \code{\link{NOVOboot}}.} \item{oda.model}{Data frame
  containing the parsed model output and the ESS, ESP, and significance for
  ODA models.} \item{oda.perf}{Data frame containing the overall accuracy,
  sensitivity, specificity, positive predictive value, and negative
  predictive value for each attribute included in the ODA model. Univariate
  OR and 95\% CI are also presented. The Haldane-Anscombe-Gart correction is
  made observed cell counts of zero and a warning is displayed.}
  \item{oda.stats}{Data frame containing the classification summary for each
  ODA model contained in the model.out file.} \item{oda.sda}{Data frame
  containing predictions based on the model attributes and observed
  classifications. Correct and incorrect classifications, e.g., False
  positive and false negative status, are captured for each observation in
  the dataset. Structural decomposition can be completed using these data in
  subsequent ODA models.}
}
\description{
Parses model details, model predictions, and loads objects.
}
\details{
When run, \code{ODAparse} will return model performance metrics and
  data loaded within the global environment, which can be further evaluated.

  The working directory must be directed toward the project files to be
  loaded and located within a \code{Run} folder inside the project tree.

  For each ODA model, \code{ODAparse} will return the \strong{Effect Strength
  for Sensitivity} or ESS and the \strong{Effect Strength for Predictive
  Value} or ESP. Mean \strong{Percent Accuracy in Classification} (PAC) and
  \strong{Mean Predicitve Value} (MPV) are reported, as these are a common
  metrics for predictive model performance.

  In binary classification problems the following relationships are defined:

  \deqn{PAC = \frac{(Sensitivity+Specificity)}{2} \times 100}

  \deqn{ESS = \frac{(PAC-50)}{(100-50)} \times 100}

  \deqn{MPV = \frac{(PPV+NPV)}{2} \times 100}

  \deqn{ESP = \frac{(MPV-50)}{(100-50)} \times 100}

  Unlike mean PAC and MPV, both ESS and ESP are normed against chance
  (Yarnold \emph{et al}. 2005 and 2016) providing an intuitive scaling of
  model accuracy within the ODA paradigm.

  The significance values reported are from 1. Monte Carlo simulations for
  all observations and 2. Fisher's Exact tests for LOO analysis. Depending on
  whether a GEN, CAT, or ordered ODA model is detected, and if LOO jackknife
  is performed, the exact \code{oda.model} output will vary.

  The user is referred to the ODA User guide \code{\link{ODAmanual}} and is
  encouraged to review the data contained within the MODEL.OUT file for more
  information.

  \code{ODAparse} is not compatible with models where mcarlo = F see
  \code{ODArun}

  Users may receive the unexpected warning: "Warning  7: WEIGHTed CATEGORICAL
  LOO not available. LOO switched to OFF." If this warning does not reflect
  specification of a categorical weighted model, it is possible that the
  attribute does not vary within one or more classes. Selection of a
  different attribute / predictor is suggested.
}
\examples{
# Not run:
# ODAparse(1)

}
\references{
Yarnold P.R. and Soltysik R.C. (2005). \emph{Optimal data
  analysis: Guidebook with software for Windows}. APA Books.

  Yarnold, P.R. and Soltysik, R.C. (2016). \emph{Maximizing Predictive
  Accuracy}. ODA Books. DOI: 10.13140/RG.2.1.1368.3286.
}
\seealso{
\code{\link{ODAmanual}} \code{\link{ODAclean}} \code{\link{ODAload}}
  \code{\link{ODArun}} \code{\link{NOVOboot}}
}
\author{
Nathaniel J. Rhodes
}
