% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODApower.R
\name{ODApower}
\alias{ODApower}
\title{Estimate power for an ODA model.}
\usage{
ODApower(n1, n2, p1, p2, comp, alpha, nsim)
}
\arguments{
\item{n1}{A numeric vector that contains the number of subjects in group 1}

\item{n2}{A numeric vector that contains the number of subjects in group 2}

\item{p1}{A numeric value that is the proportion of group 1 with the outcome}

\item{p2}{A numeric value that is the proportion of group 2 with the outcome}

\item{comp}{An integer value specifying the number of experiment wise
comparions for a Sidak type adjustment to \code{alpha}}

\item{alpha}{Numeric value specifying the a priori level of signifanced
assumed}

\item{nsim}{An integer value specifying the number of Fisher's Exact Tests to
simulate.}
}
\value{
An array of power estimates with \code{nrows} of length \code{n1} and
  \code{ncol} of length \code{comp}
}
\description{
Statistical power is estimated for a unit weighted binary
  application with balanced samples in each of 2 groups
}
\details{
A default of 10,000 Monte Carlo Fisher's Exact tests are simulated
  and compared to \code{alpha} to estimate power.

  The resulting power estimate represents a "worst case scenario" for the
  lowest level of measurement accuracy, i.e., a two class and two attribute
  problem, see Rhodes 2020.

  For unit weighted applications, a Fisher's Exact test is isomorphic to the
  power of an ODA model, see Yarnold \emph{et al.} 2005.

  The \emph{a priori} signifiance level \code{alpha} is adjusted based on
  number of comparisons (\code{comp}) as follows: \eqn{alpha_{adjusted} =
  1-(1-alpha)^{1/comp}}.
}
\examples{
n1 <- seq(15,50,5)
n2 <- seq(15,50,5)
p1 <- 0.74
p2 <- 0.26
alpha <- 0.05
comp <- 1
nsim <- 100
#Power for an analysis with an ESS of 48\% (a moderate effect)
ess <- 100*(((0.74+0.74)/2)-0.5)/0.5
ODApower(n1=n1,n2=n2,comp=comp,p1=p1,p2=p2,alpha=alpha,nsim=nsim)

}
\references{
Rhodes N.J. (2020). Statistical Power Analysis in ODA, CTA and
  Novometrics. \emph{Optimal Data Analysis} \bold{9}, 21-25.
  \url{https://odajournal.files.wordpress.com/2020/02/v9a5.pdf}

  Yarnold P.R. and Soltysik R.C. (2005). \emph{Optimal data analysis:
  Guidebook with software for Windows}. APA Books.
}
\seealso{
\code{\link{fisher.test}} \code{\link{power.fisher.test}}
}
\author{
Nathaniel J. Rhodes
}
