% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODArun.R
\name{ODArun}
\alias{ODArun}
\title{Execute an ODA model run.}
\usage{
ODArun(
  run = "",
  path = getwd(),
  data = "data.txt",
  out = "model.out",
  hold = "",
  vstart = "",
  vend = "",
  class = "",
  attribute = "",
  categorical = "",
  include = "",
  exclude = "",
  direction = "",
  degen = "",
  gen = "",
  primary = "",
  secondary = "",
  nopriors = F,
  miss = "",
  weight = "",
  mcarlo = T,
  iter = "25000",
  target = "",
  sidak = "",
  stop = "",
  adjust = F,
  setseed = "",
  loooff = F,
  overwrite = FALSE
)
}
\arguments{
\item{run}{A numerical value specifying the run folder containing the data}

\item{path}{The working directory of the project stored as \code{path}}

\item{data}{A character name specifying the data.txt file in the runs folder}

\item{out}{A character name specifying the output file with "model.out" as
the default}

\item{hold}{A character name specifying the holdout data file, omitted by
default}

\item{vstart}{A character name specifying the start variable, see \code{key}
object from \code{\link{ODAload}}}

\item{vend}{A character name specifying the end variable, see \code{key}
object from \code{\link{ODAload}}}

\item{class}{A character name specifying the class variable, see \code{key}
object from \code{\link{ODAload}}}

\item{attribute}{A character name specifying the attributes, see \code{key}
object from \code{\link{ODAload}}}

\item{categorical}{An optional character name specifying categorical
attributes, see \code{key} object from \code{\link{ODAload}}}

\item{include}{An optional character name specifying the variable and a value
of observations that are included e.g., \code{"v2=2"} or \code{"v3>=50"}}

\item{exclude}{An optional character name specifying the variable and a value
that are excluded}

\item{direction}{An optional character name and direction e.g., \code{v2 < 1
or v2 LT 1} specifying a directional hypothesis for the class variable}

\item{degen}{An optional character name specifying attributes for which
degenerate solutions are allowed, off by default}

\item{gen}{An optional character name specifying the variable whose values
denote groups for a multisample generalizablity analysis, off by default}

\item{primary}{An optional character vector specifying the primary criterion
for choosing among optimal solutions, see \code{Details}}

\item{secondary}{An optional character vector specifying the secondary
criterion for choosing among optimal solutions, see \code{Details}}

\item{nopriors}{An optional logical value specifying whether the ODA
criterion is weighted by the reciprocal of class membership, with
\code{nopriors = FALSE} as the default}

\item{miss}{A numeric value specifying a missing or \code{NA} value in the
data with a default value set at \code{-9}, see \code{\link{ODAclean}}}

\item{weight}{An optional character name of a variable containing a positive,
non zero, weight value; cannot be the same as variables declared as
\code{class} \code{attribute} \code{categorical} or \code{gen}}

\item{mcarlo}{A logical value specifying whether Monte Carlo analysis should
be used to estimate Type I error with \code{mcarlo = TRUE} as the default}

\item{iter}{An integer value specifying the maximum number of Monte Carlo
iteratations to be reached before halting and must be specified if
\code{mcarlo = TRUE}}

\item{target}{An optional numerical value specifying the target level of
\code{alpha < target} to be reached before halting and must be specified if
\code{sidak} or \code{stop} are utilized}

\item{sidak}{An optional integer value specifying an adjustment to
\code{target} based on the number of experiment wise comparisons and must
be combined with \code{target}}

\item{stop}{An optional numerical value specifying the confidence level that
Type I error is less than \code{target} to be reached before halting and
must be combined with \code{target}}

\item{adjust}{An optional logical value specifying whether tied Monte Carlo
iterations are to be split in half with \code{adjust = FALSE} as the
default}

\item{setseed}{An optional integer value specifying a seed number for Monte
Carlo analysis with the current system time as the default}

\item{loooff}{An optional logical value specifying whether leave one out or
LOO analysis should be turned off with \code{loooff = FALSE} as the default}

\item{overwrite}{An optional logical value specifying whether previous files
in the \code{Run} folder should be overwritten with \code{overwrite =
FALSE} as the default}
}
\value{
Nothing is returned. Three files are created in the \code{Run}
  folder: \item{MODEL.out}{The model file that contains the commands from
  MEGAODA syntax and analysis results, see \code{\link{ODAmanual}}. This file
  is required for \code{\link{ODAparse}}} \item{OUT.out}{The echo file that
  contains the inital commands for OPEN and OUTPUT from MEGAODA syntax, see
  \code{\link{ODAmanual}}.} \item{OUT.pgm}{The command file that contains all
  of the commands for MEGAODA passed from R based on user input to
  \code{ODArun}.}
}
\description{
Creates an command file using the parameters below and calls
  MegaODA
}
\details{
The working directory of the project is stored as \code{path}. All
  files needed for the run must be located in the appropriate \code{run}
  folder beneath this level. See \code{\link{ODAclean}}.

  \code{ODArun} will produce a command file with an extension of .pgm and two
  files with .out extensions. Resulting MODEL.OUT files can be parsed using
  \code{\link{ODAparse}}.

  The ODA algorithm explicitly maximizes the classification accuracy which is
  achieved for the training sample see Yarnold, 2005.

  The use of Optimal in Optimal Data Analysis means that an ODA model
  achieves the theoretically maximum possible level of accuracy in any given
  application. For more information see \code{\link{ODAmanual}}.

  ODA utilizes primary and secondary criteria for selecting among multiple
  optimal solutions. By default, when not specified and when \code{priorsoff
  = FALSE} \code{primary} is set to \code{maxsens}. By default, when not
  specified \code{secondary} is set to \code{samplerep}.

  When \code{gen} is not active, other options include: \code{maxsens}
  \code{meansens} \code{samplerep} \code{balanced} \code{distance}
  \code{random sens(attribute)}

  When \code{gen} is active, other options include: \code{genmean} and
  \code{gensens(attribute)}

  There are several disallowed specifications. Error checking is
  automatically performed on the user inputs. However, if the \code{to}
  keyword is used with a range of variables, it is suggested that the user
  confirm that the desired analysis was performed as some error checking is
  not available. The following cannot be combined in \code{ODArun}:
  \code{weight} cannot be both declared and listed as any of the following
  \code{class attribute gen}. Likewise, \code{gen} cannot be both declared
  and listed as any of the following \code{attribute} \code{categorical}
  \code{class} \code{weight}.
}
\examples{
# Not run:
# ODArun(run=1, vstart="v1",vend="v45",class="v45",attribute="v1 to v44", categorical="v1 v3")

}
\references{
Yarnold P.R. and Soltysik R.C. (2005). \emph{Optimal data
  analysis: Guidebook with software for Windows}. APA Books.

  Yarnold, P.R. and Soltysik, R.C. (2016). \emph{Maximizing Predictive
  Accuracy}. ODA Books. DOI: 10.13140/RG.2.1.1368.3286
}
\seealso{
\code{\link{ODAmanual}} \code{\link{ODAclean}} \code{\link{ODAload}}
}
\author{
Nathaniel J. Rhodes
}
