% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAsummary.R
\name{ODAsummary}
\alias{ODAsummary}
\title{Merges results of ODAparse into printable report.}
\usage{
ODAsummary(run = "", ...)
}
\arguments{
\item{run}{The numerical value of the run folder in which the data file is
stored}

\item{...}{Additional run numbers specified as a list}
}
\value{
The following objects are returned for ODA models:
  \item{oda.summary.X}{Merged Object containing "oda.model.X" and
  "oda.perf.X" where X is specified by the \code{run} argument.}
}
\description{
Transforms objects created by \code{ODAparse} and loads results.
}
\details{
When run, \code{ODAsummary} will merge the ODA model summary and
  model performance metrics that are stored in the global environment.

  The \code{run} number should be an existing run file within the current
  project tree. To function properly, \code{ODAparse} must first be completed
  for each run for which a summary report is being requested.

  For more information on the objects generated by \code{ODAparse} see
  documentation for \code{ODAparse}. Specifically, \code{ODAsummary} will
  merge the object "oda.model.X" and "oda.perf.X" where X is the \code{run}
  number specified as an argument to \code{ODAparse} and \code{ODAsummary}.

  The "oda.summary.X" object also includes several new variables each of
  which indicate whether the model results are significant in training
  "mcpsig" or LOO "loosig" or LOO-stable "loostab" indicating the ESS in
  training is the same as the ESS in LOO.

  For models where loooff=T, "loostab" and "loosig" are not reported, see
  \code{ODArun} for more details on LOO specifications.

  For GEN models, the report is limited to the overall model. Users should
  review the \code{ODAparse} outputs and the "oda.model.X" summary object to
  evaluate the results for each GEN sample / subgroup.
}
\examples{

#' # Not run:
# ODAsummary(1)
# print(oda.summary.1)

}
\seealso{
\code{\link{ODAmanual}} \code{\link{ODAparse}} \code{\link{ODArun}}
}
\author{
Nathaniel J. Rhodes
}
