% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAtree.R
\name{ODAtree}
\alias{ODAtree}
\title{Generate a folder tree for an ODA project}
\usage{
ODAtree(project = "NewProject", folder = getwd())
}
\arguments{
\item{project}{A character string of a new project, e.g. "New Project"}

\item{folder}{The full path to the root folder for the new project. Default
is the current working directory.}
}
\value{
A new folder named as \code{project} containing the following
  subfolders: \item{Rscript}{The folder for the Rscript file with templated
  syntax as a skeleton Rscript for the project.} \item{Runs}{The folder for
  the \code{\link{ODArun}} outputs and data files for cleaning and analysis.}
  \item{Program}{The folder containing the executable program used for all
  \code{\link{ODArun}} analyses.}
}
\description{
Establishes a subdirectory for an ODA project within a given working
directory.
}
\examples{
##Not run
##ODAtree("NewProject")
}
