% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAclean.R
\name{ODAclean}
\alias{ODAclean}
\title{Read and clean a data.csv input file and transform variables for use
  in \code{ODArun} or \code{CTArun}}
\usage{
ODAclean(
  data = "",
  output = "",
  type = "",
  miss = "",
  ipsative = "",
  normative = "",
  id = "",
  overwrite = FALSE
)
}
\arguments{
\item{data}{The character name of the .csv file to be loaded and cleaned. The
current working directory must be set to the Project folder.}

\item{output}{An integer that specifies of the folder into which the cleaned
data files will be written. If the directory does not exist, it will be
created. If it does exist, the user will be asked whether the file contents
should be overwritten.}

\item{type}{A character string which should contain either "ODA" or "CTA".
Default value is "ODA".}

\item{miss}{A numeric value e.g., -9 that is imputed for all \code{NA} values
in the imported data frame where missing or \code{NA} values exist.}

\item{ipsative}{A character vector of variable names \code{x} in the data
which will be standardized within an \code{id} i.e., \eqn{x - mean(x) /
sd(x)}. An \code{id} variable must be supplied.}

\item{normative}{A character vector of variable names \code{x} in the cleaned
data which will be standardized across columns \eqn{x - mean(x) / sd(x)}.}

\item{id}{A character vector that represents repeated measures. Observations
within an \code{id} are standardized to the mean of the \code{id} number.
More than 1 observation per subject is required.}

\item{overwrite}{Logical value specifying whether files in output directory
should be overwritten. Can be overridden by specifying \code{TRUE} or
\code{FALSE}.}
}
\value{
Cleaned data is moved to the model type folder as both .txt and .csv
  files. \item{data.txt}{A cleaned data file is moved to the output
  directory. Row and column names are removed from the .txt file. If
  specified, missing value replacements and standardized variables are also
  written to this file.} \item{data.csv}{The .csv file is also moved to the
  output directory.}
}
\description{
A valid .csv file is imported, cleaned, and moved to a specified
  output folder. Data frame objects called \code{key} and \code{data} are
  loaded in the environment. The .csv file should be one level above the
  output folder.
}
\examples{
# Not Run
# ODAclean(data="data.csv",type='ODA',output=1, miss=-9)
}
\author{
Nathaniel J. Rhodes
}
