% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAtree.R
\name{ODAtree}
\alias{ODAtree}
\title{Generate a folder tree for an ODA project}
\usage{
ODAtree(project = "NewProject", folder = getwd())
}
\arguments{
\item{project}{A character string of a new project, e.g. "New Project"}

\item{folder}{The full path to the root folder for the new project. Default
is the current working directory.}
}
\value{
A new folder named as \code{project} containing the following
  subfolders:\item{CTA}{The folder to contain the \code{\link{CTArun}}
  outputs and data files for the project.} \item{ODA}{The folder to contain
  the \code{\link{ODArun}} outputs and data files for the project.}
  \item{Program}{The folder containing the executable programs used for
  \code{\link{ODArun}} and \code{\link{CTArun}} analyses.} \item{Rscript}{The
  folder for the Rscript file with a skeleton Rscript containing command
  syntax for the project.}
}
\description{
Establishes a subdirectory for an ODA project within a given working
directory.
}
\examples{
##Not run
##ODAtree("NewProject")
}
